requirements.txt
selenium
Pillow
discord.py
flask